+++ papermill/tests/test_cli.py
@@ -3,6 +3,10 @@
 """ Test the command line interface """
 
 import os
+import sys
+import subprocess
+import nbformat
+from jupyter_client import kernelspec
 import unittest
 
 import pytest
@@ -88,6 +92,29 @@ class TestCLI(unittest.TestCase):
             'output.ipynb',
             {'foo': 'bar', 'baz': 42},
             engine_name=None,
+            request_save_on_cell_execute=True,
+            prepare_only=False,
+            kernel_name=None,
+            log_output=False,
+            progress_bar=True,
+            start_timeout=60,
+            report_mode=False,
+            cwd=None,
+        )
+
+    @patch(cli.__name__ + '.execute_notebook')
+    def test_parameters_list_args(self, execute_patch):
+        self.runner.invoke(
+            papermill, self.default_args + ['-p', 'foo', 'bar', 'baz', '0',
+                                            '-p', 'one', 'two',
+                                            '--parameters', 'qux', '42']
+        )
+        execute_patch.assert_called_with(
+            'input.ipynb',
+            'output.ipynb',
+            {'foo': ['bar', 'baz', 0], 'one': 'two', 'qux': 42},
+            engine_name=None,
+            request_save_on_cell_execute=True,
             prepare_only=False,
             kernel_name=None,
             log_output=False,
@@ -107,6 +134,29 @@ class TestCLI(unittest.TestCase):
             'output.ipynb',
             {'foo': 'bar', 'baz': '42'},
             engine_name=None,
+            request_save_on_cell_execute=True,
+            prepare_only=False,
+            kernel_name=None,
+            log_output=False,
+            progress_bar=True,
+            start_timeout=60,
+            report_mode=False,
+            cwd=None,
+        )
+
+    @patch(cli.__name__ + '.execute_notebook')
+    def test_parameters_raw_list_args(self, execute_patch):
+        self.runner.invoke(
+            papermill, self.default_args + ['-r', 'foo', 'bar', 'baz', '0',
+                                            '-r', 'one', 'two',
+                                            '--parameters_raw', 'qux', '42']
+        )
+        execute_patch.assert_called_with(
+            'input.ipynb',
+            'output.ipynb',
+            {'foo': ['bar', 'baz', '0'], 'one': 'two', 'qux': '42'},
+            engine_name=None,
+            request_save_on_cell_execute=True,
             prepare_only=False,
             kernel_name=None,
             log_output=False,
@@ -127,8 +177,9 @@ class TestCLI(unittest.TestCase):
             'input.ipynb',
             'output.ipynb',
             # Last input wins dict update
-            {'foo': 54321, 'bar': 'value', 'baz': {'k2': 'v2', 'k1': 'v1'}},
+            {'foo': 54321, 'bar': 'value', 'baz': {'k2': 'v2', 'k1': 'v1'}, 'a_date': '2019-01-01'},
             engine_name=None,
+            request_save_on_cell_execute=True,
             prepare_only=False,
             kernel_name=None,
             log_output=False,
@@ -149,6 +200,28 @@ class TestCLI(unittest.TestCase):
             'output.ipynb',
             {'foo': 'bar', 'foo2': ['baz']},
             engine_name=None,
+            request_save_on_cell_execute=True,
+            prepare_only=False,
+            kernel_name=None,
+            log_output=False,
+            progress_bar=True,
+            start_timeout=60,
+            report_mode=False,
+            cwd=None,
+        )
+
+    @patch(cli.__name__ + '.execute_notebook')
+    def test_parameters_yaml_date(self, execute_patch):
+        self.runner.invoke(
+            papermill,
+            self.default_args + ['-y', 'a_date: 2019-01-01'],
+        )
+        execute_patch.assert_called_with(
+            'input.ipynb',
+            'output.ipynb',
+            {'a_date': '2019-01-01'},
+            engine_name=None,
+            request_save_on_cell_execute=True,
             prepare_only=False,
             kernel_name=None,
             log_output=False,
@@ -170,6 +243,7 @@ class TestCLI(unittest.TestCase):
             # Last input wins dict update
             {'foo': ['baz']},
             engine_name=None,
+            request_save_on_cell_execute=True,
             prepare_only=False,
             kernel_name=None,
             log_output=False,
@@ -197,6 +271,32 @@ class TestCLI(unittest.TestCase):
             # Last input wins dict update
             {'foo': 1, 'bar': 2},
             engine_name=None,
+            request_save_on_cell_execute=True,
+            prepare_only=False,
+            kernel_name=None,
+            log_output=False,
+            progress_bar=True,
+            start_timeout=60,
+            report_mode=False,
+            cwd=None,
+        )
+
+    @patch(cli.__name__ + '.execute_notebook')
+    def test_parameters_base64_date(self, execute_patch):
+        self.runner.invoke(
+            papermill,
+            self.default_args
+            + [
+                '--parameters_base64',
+                'YV9kYXRlOiAyMDE5LTAxLTAx',
+            ],
+        )
+        execute_patch.assert_called_with(
+            'input.ipynb',
+            'output.ipynb',
+            {'a_date': '2019-01-01'},
+            engine_name=None,
+            request_save_on_cell_execute=True,
             prepare_only=False,
             kernel_name=None,
             log_output=False,
@@ -215,6 +315,7 @@ class TestCLI(unittest.TestCase):
             # Last input wins dict update
             {'PAPERMILL_INPUT_PATH': 'input.ipynb'},
             engine_name=None,
+            request_save_on_cell_execute=True,
             prepare_only=False,
             kernel_name=None,
             log_output=False,
@@ -233,6 +334,7 @@ class TestCLI(unittest.TestCase):
             # Last input wins dict update
             {'PAPERMILL_OUTPUT_PATH': 'output.ipynb'},
             engine_name=None,
+            request_save_on_cell_execute=True,
             prepare_only=False,
             kernel_name=None,
             log_output=False,
@@ -251,6 +353,7 @@ class TestCLI(unittest.TestCase):
             # Last input wins dict update
             {'PAPERMILL_INPUT_PATH': 'input.ipynb', 'PAPERMILL_OUTPUT_PATH': 'output.ipynb'},
             engine_name=None,
+            request_save_on_cell_execute=True,
             prepare_only=False,
             kernel_name=None,
             log_output=False,
@@ -268,6 +371,7 @@ class TestCLI(unittest.TestCase):
             'output.ipynb',
             {},
             engine_name='engine-that-could',
+            request_save_on_cell_execute=True,
             prepare_only=False,
             kernel_name=None,
             log_output=False,
@@ -285,6 +389,7 @@ class TestCLI(unittest.TestCase):
             'output.ipynb',
             {},
             engine_name=None,
+            request_save_on_cell_execute=True,
             prepare_only=True,
             kernel_name=None,
             log_output=False,
@@ -302,6 +407,7 @@ class TestCLI(unittest.TestCase):
             'output.ipynb',
             {},
             engine_name=None,
+            request_save_on_cell_execute=True,
             prepare_only=False,
             kernel_name='python3',
             log_output=False,
@@ -319,6 +425,7 @@ class TestCLI(unittest.TestCase):
             'output.ipynb',
             {},
             engine_name=None,
+            request_save_on_cell_execute=True,
             prepare_only=False,
             kernel_name=None,
             log_output=False,
@@ -336,6 +443,7 @@ class TestCLI(unittest.TestCase):
             'output.ipynb',
             {},
             engine_name=None,
+            request_save_on_cell_execute=True,
             prepare_only=False,
             kernel_name=None,
             log_output=False,
@@ -353,6 +461,7 @@ class TestCLI(unittest.TestCase):
             'output.ipynb',
             {},
             engine_name=None,
+            request_save_on_cell_execute=True,
             prepare_only=False,
             kernel_name=None,
             log_output=False,
@@ -370,6 +479,7 @@ class TestCLI(unittest.TestCase):
             'output.ipynb',
             {},
             engine_name=None,
+            request_save_on_cell_execute=True,
             prepare_only=False,
             kernel_name=None,
             log_output=True,
@@ -387,6 +497,7 @@ class TestCLI(unittest.TestCase):
             'output.ipynb',
             {},
             engine_name=None,
+            request_save_on_cell_execute=True,
             prepare_only=False,
             kernel_name=None,
             log_output=True,
@@ -404,6 +515,7 @@ class TestCLI(unittest.TestCase):
             'output.ipynb',
             {},
             engine_name=None,
+            request_save_on_cell_execute=True,
             prepare_only=False,
             kernel_name=None,
             log_output=False,
@@ -421,6 +533,7 @@ class TestCLI(unittest.TestCase):
             'output.ipynb',
             {},
             engine_name=None,
+            request_save_on_cell_execute=True,
             prepare_only=False,
             kernel_name=None,
             log_output=False,
@@ -438,6 +551,7 @@ class TestCLI(unittest.TestCase):
             'output.ipynb',
             {},
             engine_name=None,
+            request_save_on_cell_execute=True,
             prepare_only=False,
             kernel_name=None,
             log_output=False,
@@ -455,6 +569,7 @@ class TestCLI(unittest.TestCase):
             'output.ipynb',
             {},
             engine_name=None,
+            request_save_on_cell_execute=True,
             prepare_only=False,
             kernel_name=None,
             log_output=False,
@@ -472,6 +587,7 @@ class TestCLI(unittest.TestCase):
             'output.ipynb',
             {},
             engine_name=None,
+            request_save_on_cell_execute=True,
             prepare_only=False,
             kernel_name=None,
             log_output=False,
@@ -525,8 +641,10 @@ class TestCLI(unittest.TestCase):
                 'baz': 'replace',
                 'yaml_foo': {'yaml_bar': 'yaml_baz'},
                 "base64_foo": "base64_bar",
+                'a_date': '2019-01-01',
             },
             engine_name='engine-that-could',
+            request_save_on_cell_execute=True,
             prepare_only=True,
             kernel_name='R',
             log_output=True,
@@ -571,3 +689,129 @@ def test_papermill_log():
     assert result.exit_code == 0
     assert 'Log output mode is on' in result.output
     assert 'Logging' in result.output
+
+
+def papermill_cli(papermill_args=None, **kwargs):
+    cmd = [sys.executable, '-m', 'papermill.cli']
+    if papermill_args:
+        cmd.extend(papermill_args)
+    return subprocess.Popen(cmd, **kwargs)
+
+
+def papermill_version():
+    try:
+        proc = papermill_cli(['--version'], stdout=subprocess.PIPE)
+        out, _ = proc.communicate()
+        if proc.returncode:
+            return None
+        return out.decode('utf-8')
+    except (OSError, SystemExit):  # pragma: no cover
+        return None
+
+
+@pytest.fixture()
+def notebook():
+    for name in kernelspec.find_kernel_specs():
+        ks = kernelspec.get_kernel_spec(name)
+        metadata = {'kernelspec': {'name': name,
+                                   'language': ks.language,
+                                   'display_name': ks.display_name}}
+        return nbformat.v4.new_notebook(
+            metadata=metadata,
+            cells=[nbformat.v4.new_markdown_cell(
+                'This is a notebook with kernel: ' + ks.display_name)])
+
+    raise EnvironmentError('No kernel found')
+
+
+require_papermill_installed = pytest.mark.skipif(
+    not papermill_version(),
+    reason='papermill is not installed')
+
+
+@require_papermill_installed
+def test_pipe_in_out_auto(notebook):
+    process = papermill_cli(stdout=subprocess.PIPE, stdin=subprocess.PIPE)
+    text = nbformat.writes(notebook)
+    out, err = process.communicate(input=text.encode('utf-8'))
+
+    # Test no message on std error
+    assert not err
+
+    # Test that output is a valid notebook
+    nbformat.reads(out.decode('utf-8'), as_version=4)
+
+
+@require_papermill_installed
+def test_pipe_in_out_explicit(notebook):
+    process = papermill_cli(['-', '-'], stdout=subprocess.PIPE, stdin=subprocess.PIPE)
+    text = nbformat.writes(notebook)
+    out, err = process.communicate(input=text.encode('utf-8'))
+
+    # Test no message on std error
+    assert not err
+
+    # Test that output is a valid notebook
+    nbformat.reads(out.decode('utf-8'), as_version=4)
+
+
+@require_papermill_installed
+def test_pipe_out_auto(tmpdir, notebook):
+    nb_file = tmpdir.join('notebook.ipynb')
+    nb_file.write(nbformat.writes(notebook))
+
+    process = papermill_cli([str(nb_file)], stdout=subprocess.PIPE)
+    out, err = process.communicate()
+
+    # Test no message on std error
+    assert not err
+
+    # Test that output is a valid notebook
+    nbformat.reads(out.decode('utf-8'), as_version=4)
+
+
+@require_papermill_installed
+def test_pipe_out_explicit(tmpdir, notebook):
+    nb_file = tmpdir.join('notebook.ipynb')
+    nb_file.write(nbformat.writes(notebook))
+
+    process = papermill_cli([str(nb_file), '-'], stdout=subprocess.PIPE)
+    out, err = process.communicate()
+
+    # Test no message on std error
+    assert not err
+
+    # Test that output is a valid notebook
+    nbformat.reads(out.decode('utf-8'), as_version=4)
+
+
+@require_papermill_installed
+def test_pipe_in_auto(tmpdir, notebook):
+    nb_file = tmpdir.join('notebook.ipynb')
+
+    process = papermill_cli([str(nb_file)], stdin=subprocess.PIPE)
+    text = nbformat.writes(notebook)
+    out, _ = process.communicate(input=text.encode('utf-8'))
+
+    # Nothing on stdout
+    assert not out
+
+    # Test that output is a valid notebook
+    with open(str(nb_file)) as fp:
+        nbformat.reads(fp.read(), as_version=4)
+
+
+@require_papermill_installed
+def test_pipe_in_explicit(tmpdir, notebook):
+    nb_file = tmpdir.join('notebook.ipynb')
+
+    process = papermill_cli(['-', str(nb_file)], stdin=subprocess.PIPE)
+    text = nbformat.writes(notebook)
+    out, _ = process.communicate(input=text.encode('utf-8'))
+
+    # Nothing on stdout
+    assert not out
+
+    # Test that output is a valid notebook
+    with open(str(nb_file)) as fp:
+        nbformat.reads(fp.read(), as_version=4)